# docs/openapi-v1.json
{
  "components": {
    "schemas": {
      "AnalyticsFilters": {
        "properties": {
          "anio": {
            "items": {
              "type": "string"
            },
            "title": "Anio",
            "type": "array"
          },
          "categoria": {
            "items": {
              "type": "string"
            },
            "title": "Categoria",
            "type": "array"
          },
          "contract_month": {
            "items": {
              "type": "string"
            },
            "title": "Contract Month",
            "type": "array"
          },
          "gestion_month": {
            "items": {
              "type": "string"
            },
            "title": "Gestion Month",
            "type": "array"
          },
          "supervisor": {
            "items": {
              "type": "string"
            },
            "title": "Supervisor",
            "type": "array"
          },
          "tramo": {
            "items": {
              "type": "string"
            },
            "title": "Tramo",
            "type": "array"
          },
          "un": {
            "items": {
              "type": "string"
            },
            "title": "Un",
            "type": "array"
          },
          "via_cobro": {
            "items": {
              "type": "string"
            },
            "title": "Via Cobro",
            "type": "array"
          },
          "via_pago": {
            "items": {
              "type": "string"
            },
            "title": "Via Pago",
            "type": "array"
          }
        },
        "title": "AnalyticsFilters",
        "type": "object"
      },
      "ExportRequest": {
        "properties": {
          "endpoint": {
            "maxLength": 64,
            "minLength": 3,
            "title": "Endpoint",
            "type": "string"
          },
          "filters": {
            "$ref": "#/components/schemas/AnalyticsFilters"
          },
          "format": {
            "pattern": "^(csv|pdf)$",
            "title": "Format",
            "type": "string"
          }
        },
        "required": [
          "format",
          "endpoint"
        ],
        "title": "ExportRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginIn": {
        "properties": {
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
      "RefreshIn": {
        "properties": {
          "refresh_token": {
            "maxLength": 1024,
            "minLength": 10,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshIn",
        "type": "object"
      },
      "RevokeIn": {
        "properties": {
          "refresh_token": {
            "maxLength": 1024,
            "minLength": 10,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RevokeIn",
        "type": "object"
      },
      "RulesIn": {
        "properties": {
          "rules": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rules",
            "type": "array"
          }
        },
        "title": "RulesIn",
        "type": "object"
      },
      "RulesOut": {
        "properties": {
          "rules": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rules",
            "type": "array"
          }
        },
        "title": "RulesOut",
        "type": "object"
      },
      "SupervisorsScopeIn": {
        "properties": {
          "supervisors": {
            "items": {
              "type": "string"
            },
            "title": "Supervisors",
            "type": "array"
          }
        },
        "title": "SupervisorsScopeIn",
        "type": "object"
      },
      "SupervisorsScopeOut": {
        "properties": {
          "supervisors": {
            "items": {
              "type": "string"
            },
            "title": "Supervisors",
            "type": "array"
          }
        },
        "title": "SupervisorsScopeOut",
        "type": "object"
      },
      "TokenOut": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "role",
          "permissions"
        ],
        "title": "TokenOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "cobranzas-api-v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analytics/brokers/summary": {
      "post": {
        "operationId": "brokers_summary_api_v1_analytics_brokers_summary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsFilters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Brokers Summary",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/export": {
      "post": {
        "operationId": "analytics_export_api_v1_analytics_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Analytics Export",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/mora/summary": {
      "post": {
        "operationId": "mora_summary_api_v1_analytics_mora_summary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsFilters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mora Summary",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/portfolio/summary": {
      "post": {
        "operationId": "portfolio_summary_api_v1_analytics_portfolio_summary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsFilters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Portfolio Summary",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/rendimiento/summary": {
      "post": {
        "operationId": "rendimiento_summary_api_v1_analytics_rendimiento_summary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsFilters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rendimiento Summary",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/revoke": {
      "post": {
        "operationId": "revoke_api_v1_auth_revoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/brokers/commissions": {
      "get": {
        "operationId": "get_commissions_api_v1_brokers_commissions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulesOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Commissions",
        "tags": [
          "brokers"
        ]
      },
      "post": {
        "operationId": "save_commissions_api_v1_brokers_commissions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Commissions",
        "tags": [
          "brokers"
        ]
      }
    },
    "/api/v1/brokers/prizes": {
      "get": {
        "operationId": "get_prizes_api_v1_brokers_prizes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulesOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Prizes",
        "tags": [
          "brokers"
        ]
      },
      "post": {
        "operationId": "save_prizes_api_v1_brokers_prizes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Prizes",
        "tags": [
          "brokers"
        ]
      }
    },
    "/api/v1/brokers/supervisors-scope": {
      "get": {
        "operationId": "get_supervisors_scope_api_v1_brokers_supervisors_scope_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupervisorsScopeOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Supervisors Scope",
        "tags": [
          "brokers"
        ]
      },
      "post": {
        "operationId": "save_supervisors_scope_api_v1_brokers_supervisors_scope_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupervisorsScopeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupervisorsScopeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Supervisors Scope",
        "tags": [
          "brokers"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    }
  }
}